shopi_api/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-dotenv==1.4.2
djangorestframework==3.16.1
psycopg2==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
